requirements.txt
flask
google-generativeai 
python-dotenv
python-docx
pypdf2
gunicorn